updated Learner interface added enums for gridtypes and SLE solvers fixed conversions warnings
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2078 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/datatypes/DataVectorSP.cpp
@@ -257,9 +257,9 @@ float DataVectorSP::maxNorm() {
     float max = 0.0f;
     for (size_t i = 0; i < size; i++)
     {
-        if (max < fabs(data[i]))
+        if (max < fabs((float)data[i]))
         {
-        	max = fabs(data[i]);
+        	max = fabs((float)data[i]);
         }
     }
     return max;
